# KAFKA-17879: test_performance_services.py should use DEV version to run kafka service