@chia7712 2024-10-27
the failed e2e will be fixed in follow-up (https://issues.apache.org/jira/browse/KAFKA-17883)